requirements.txt
boto3==1.5.27
moto==1.2.0
psycopg2==2.7.4
pycrypto==2.6.1